Add Curios Support for Tome of Peritia
* Tome of Peritial added as (another) Charm.  BM still only registers 1 Charm Slot.
* Moved the `getAllInventories` method out of PlayerDemonWillHandler class into its own Inventory Helper class.  Adjusted remainder of the PlayerDemonWillHandler class to use the new method location.
* Adjusted GenericHandler to use the new Inventory Handler.
