+++ src/main/resources/data/curios/tags/items/charm.json
@@ -5,6 +5,7 @@
     "bloodmagic:growthsigil",
     "bloodmagic:sigilofmagnetism",
     "bloodmagic:icesigil",
-    "bloodmagic:sigilofholding"
+    "bloodmagic:sigilofholding",
+    "bloodmagic:experiencebook"
   ]
 }
